requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blip-ci==0.0.5
brisque==0.0.15
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clip-interrogator==0.6.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
fairscale==0.4.4
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.2
firebase-admin==6.2.0
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-generativeai==0.1.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gradio==3.35.2
gradio_client==0.2.7
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
image-quality==1.2.7
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.2
libsvm==3.23.0.4
lightning-utilities==0.8.0
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mplcursors==0.5.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
orjson==3.9.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==1.7.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.11.0
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1+cu117
torchmetrics==0.11.4
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
yarl==1.9.2
zipp==3.15.0
